Deletion Test Passing, new safetyCircle

+++ app/src/main/java/com/example/locationtest/MapsActivity.java
@@ -10,6 +10,7 @@ import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
+import android.graphics.Color;
 import android.location.Address;
 import android.location.Geocoder;
 import android.location.Location;
@@ -29,6 +30,8 @@ import com.google.android.gms.maps.OnMapReadyCallback;
 import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.UiSettings;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
+import com.google.android.gms.maps.model.Circle;
+import com.google.android.gms.maps.model.CircleOptions;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
@@ -50,7 +53,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
     private GoogleMap mMap;
     private static final String BASE_URL = "http://52.149.135.175:80";
     private final static String TAG = "MapActivity";
-    private final static double nearbyDistance = 8; //km
+    private final static double nearbyDistance = 4; //km
     public List<Incident> incidents = new LinkedList<>();
     private LatLng myLocation;
     private Marker blueMarker;
@@ -58,6 +61,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
     private Incident selectedIncident;
     private boolean incidentSelected = false;
     private String deletePassword;
+    private Circle safetyCircle;
 
 
     @SuppressLint("MissingPermission")
@@ -110,6 +114,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         mMap.setOnMapClickListener(new GoogleMap.OnMapClickListener() {
             @Override
             public void onMapClick(LatLng latLng) {
+                safetyCircle.remove();
                 blueMarker.setPosition(latLng);
                 incidentSelected = false;
             }
@@ -117,6 +122,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         mMap.setOnMarkerClickListener(new GoogleMap.OnMarkerClickListener() {
             @Override
             public boolean onMarkerClick(Marker marker) {
+                safetyCircle.remove();
                 String markerTitle = marker.getTitle();
                 incidentSelected = !markerTitle.equals("You are here");
                 String incidentTitle = markerTitle.split(" -> Severity: ", 2)[0];
@@ -137,17 +143,18 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         mMap.setOnMyLocationButtonClickListener(new GoogleMap.OnMyLocationButtonClickListener() {
             @Override
             public boolean onMyLocationButtonClick() {
+                safetyCircle.remove();
                 blueMarker.setPosition(myLocation);
                 return false;
             }
         });
 
-        // Wait 5 seconds before displaying current location
-        Toast.makeText(MapsActivity.this, "Give us a moment...", Toast.LENGTH_LONG).show();
+        // Wait 1.5 seconds before displaying current location
         (new Handler()).postDelayed(this::displayCurrentLocation, 1500);
 
         // Call the sever and get the list of incidents
         retrieveIncidents();
+
     }
 
     private void retrieveIncidents() {
@@ -169,7 +176,6 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
 
                     mMap.addMarker(new MarkerOptions().position(i.getLocation()).title(i.getTitle() + " -> Severity: " + i.getSeverity()));
                 }
-                Toast.makeText(MapsActivity.this, "Response was successful", Toast.LENGTH_LONG).show();
             }
 
             @Override
@@ -197,6 +203,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
             @Override
             public void onClick(View v) {
                 try {
+                    safetyCircle.remove();
                     computeSafetyScore();
                 } catch (IOException e) {
                     Toast.makeText(MapsActivity.this, "Error", Toast.LENGTH_LONG).show();
@@ -221,22 +228,9 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         deleteIncidentButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-
-                if (incidentSelected) {
-                    DeleteDialog deleteDialog = new DeleteDialog();
-                    deleteDialog.show(getSupportFragmentManager(), "delete dialog");
-                }
-                else{
-                    Toast.makeText(MapsActivity.this, "No Incident Selected", Toast.LENGTH_LONG).show();
-                    return;
-                }
-
-
-
-
-
-
-
+                safetyCircle.remove();
+                DeleteDialog deleteDialog = new DeleteDialog();
+                deleteDialog.show(getSupportFragmentManager(), "delete dialog");
             }
         });
     }
@@ -267,6 +261,8 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
     // Calculate the safety score at the blue marker based on number of nearby incidents
     private void computeSafetyScore() throws IOException {
 
+        displayNearbyRadius();
+
         Call<ScoreMessage> call = retrofitInterface.getSafetyScore(String.valueOf((float) blueMarker.getPosition().latitude), String.valueOf((float)blueMarker.getPosition().longitude));
         call.enqueue(new Callback<ScoreMessage>() {
             @Override
@@ -281,59 +277,16 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
             }
         });
 
+    }
 
-
-/*
-        int score = 0;
-        String isSafe = "";
-        int nearbyIncidents = 0;
-        boolean geocoderWorking = true;
-
-
-        // Count the number of nearby incidents
-        for (Incident i : incidents) {
-            boolean sameLocality = true;
-
-            // If the geocoder is working, use it to determine if the incident is in the same
-            // locality as the user's current location
-            if (geocoderWorking) {
-                try {
-                    Geocoder coder = new Geocoder(this, Locale.getDefault());
-                    sameLocality = i.sameLocality(myLocation, coder);
-                } catch (IOException e) {
-                    geocoderWorking = false;
-                }
-            }
-
-            // If the incident is nearby, increment nearby incidents
-            if (sameLocality) {
-                if (i.distanceFrom(blueMarker.getPosition()) < nearbyDistance) {
-                    nearbyIncidents += 1;
-                }
-            }
-        }
-
-        // Score from 1 to 5
-        // Many incidents -> Low score
-        // No incidents -> High score
-        if (nearbyIncidents >= 10) {
-            score = 1;
-            isSafe = "(Unsafe)";
-        } else if (nearbyIncidents >= 7) {
-            score = 2;
-        } else if (nearbyIncidents >= 4) {
-            score = 3;
-        } else if (nearbyIncidents >= 1) {
-            score = 4;
-            isSafe = "(Safe)";
-        } else {
-            score = 5;
-            isSafe = "(Very Safe)";
-        }
-
-        // Show the safety score to the user
-        String toPrint = "The safety score at this location is: " + score + "/5 " + isSafe;
-        Toast.makeText(MapsActivity.this, toPrint, Toast.LENGTH_LONG).show();*/
+    private void displayNearbyRadius() {
+        CircleOptions options = new CircleOptions()
+                .center(blueMarker.getPosition())
+                .radius(1000*nearbyDistance)
+                .fillColor(0x33FF0000)
+                .strokeColor(Color.BLACK)
+                .strokeWidth(3);
+        safetyCircle = mMap.addCircle(options);
     }
 
 
@@ -343,6 +296,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         mMap.setOnMarkerDragListener(new GoogleMap.OnMarkerDragListener() {
             @Override
             public void onMarkerDragStart(Marker marker) {
+                safetyCircle.remove();
                 blueMarker.setIcon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_AZURE));
             }
 
@@ -358,6 +312,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
         });
         mMap.moveCamera(CameraUpdateFactory.newLatLng(myLocation));
         mMap.animateCamera(CameraUpdateFactory.newLatLngZoom(myLocation, 12.0f));
+        displayNearbyRadius();
     }
 
 
@@ -371,7 +326,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
     @Override
     public void applyTexts(String password) {
         deletePassword = password;
-        if (deletePassword.equals("Cpen321")){
+        if (deletePassword.equals("Cpen321") && incidentSelected){
             Call<Void> call = retrofitInterface.deleteIncident(selectedIncident);
             call.enqueue(new Callback<Void>() {
                 @Override
@@ -390,8 +345,10 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
 
                 }
             });
-        } else {
+        } else if (incidentSelected) {
             Toast.makeText(MapsActivity.this, "Incorrect Password", Toast.LENGTH_LONG).show();
+        } else {
+            Toast.makeText(MapsActivity.this, "No Incident Selected", Toast.LENGTH_LONG).show();
         }
     }
 }
